requirements.txt
hidapi
aenum
